Created Player Controls and Adjusted Camera Settings.
I put together some code for moving the robot around. It now moves nice and smoothly. I also wrote some code for rotating the robot left and right using the Q & E keys so that we can test it much more easily. So the robot now moves smoothly around 360°.

I added this asset I made in MagicaVoxel which is just some walls and a doorway to walk through. It's very tricky to tell how big to make the assets in MagicaVoxel but I think I made the doorway 40 units wide in MV. When I loaded it in it was too small but I doubled the size and it was about perfectly sized so I think 80 MV units lines up nicely with the robot. I coloured it black to make it stick out against the background. I messed with the camera very slightly just so that it shows a bit of the robots tracks so that you can see where the edges of the robot are which helps with not bumping into things.

The robot collides with the walls reasonably well. It is quite easy to make the robot glitch through the walls and barrel roll onto its roof if you try hard enough though so we will have to find a fix for that. Could be something simple like setting transform to 0 if it detects that it's colliding.